Change EmotionsButton design
