+++ lib/app/core/constants/constants.dart
@@ -54,3 +54,45 @@ const kSituationGradient = [
   Color.fromARGB(255, 223, 169, 8),
   Color.fromARGB(255, 199, 151, 7),
 ];
+
+List<Color> getEmotionColorGradient(int emotionId) {
+  switch (emotionId) {
+    case 1:
+      return kSurpriseGradient;
+    case 2:
+      return kAngryGradient;
+    case 3:
+      return kAgreeGradient;
+    case 4:
+      return kDissGradient;
+    case 5:
+      return kHappyGradient;
+    case 6:
+      return kBoredGradient;
+    case 7:
+      return kResiGradient;
+    default:
+      return kRedGradient;
+  }
+}
+
+Color getEmotionColor(int emotionId) {
+  switch (emotionId) {
+    case 1:
+      return const Color.fromARGB(201, 148, 11, 153);
+    case 2:
+      return const Color.fromARGB(250, 141, 3, 17);
+    case 3:
+      return const Color.fromARGB(255, 8, 136, 25);
+    case 4:
+      return const Color.fromARGB(179, 8, 77, 141);
+    case 5:
+      return const Color.fromARGB(255, 87, 150, 36);
+    case 6:
+      return const Color.fromARGB(255, 134, 34, 8);
+    case 7:
+      return const Color.fromARGB(255, 41, 51, 87);
+    default:
+      return Colors.black;
+  }
+}
